// src/lib/reward-api/models/data_FinancePaymentVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type data_FinancePaymentVO = {
    amount?: string;
    created_at?: string;
    finance_doc_id?: string;
    offsetted_amount?: string;
    offsetting_amount?: string;
    payment_address?: string;
    payment_id?: string;
    payment_status?: string;
    refunded_amount?: string;
    refunding_amount?: string;
    updated_at?: string;
};