// src/lib/reward-api/models/data_IssueRequestVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type data_IssueRequestVO = {
    amount?: string;
    created_at?: string;
    creator?: string;
    expense_type?: string;
    id?: number;
    project_id?: number;
    remark?: string;
    request_status?: string;
    unit?: string;
    updated_at?: string;
    voucher_type?: string;
};